chore: consolidate all utils
