+++ src/components/calcite-block-section/calcite-block-section.tsx
@@ -1,7 +1,7 @@
 import { Component, Element, Event, EventEmitter, Prop, h, VNode } from "@stencil/core";
 
-import { getElementDir } from "../utils/dom";
-import { CSS_UTILITY } from "../utils/resources";
+import { getElementDir } from "../../utils/dom";
+import { CSS_UTILITY } from "../../utils/resources";
 import { CSS, ICONS, TEXT } from "./resources";
 import { guid } from "../utils/guid";
 import { CalciteBlockSectionToggleDisplay } from "../interfaces";
