--- src/components/utils/dom.ts
@@ -1,80 +1,0 @@
-import { CalciteTheme } from "../interfaces";
-
-export function getElementDir(el: HTMLElement): "ltr" | "rtl" {
-  return getElementProp(el, "dir", "ltr");
-}
-
-export function getElementTheme(el: HTMLElement): CalciteTheme {
-  return getElementProp(el, "theme", "light");
-}
-
-export function getElementProp(el: HTMLElement, prop, value): any {
-  const closestWithProp = el.closest(`[${prop}]`);
-  return closestWithProp ? closestWithProp.getAttribute(prop) : value;
-}
-
-export interface CalciteFocusableElement extends HTMLElement {
-  setFocus?: () => void;
-}
-
-export function focusElement(el: CalciteFocusableElement): void {
-  if (!el) {
-    return;
-  }
-
-  "setFocus" in el && typeof el.setFocus === "function" ? el.setFocus() : el.focus();
-}
-
-interface GetSlottedOptions {
-  all?: boolean;
-  direct?: boolean;
-  selector?: string;
-}
-
-export function getSlotted<T extends Element = Element>(
-  element: Element,
-  slotName: string,
-  options: GetSlottedOptions & { all: true }
-): T[];
-export function getSlotted<T extends Element = Element>(
-  element: Element,
-  slotName: string,
-  options?: GetSlottedOptions
-): T | null;
-export function getSlotted<T extends Element = Element>(
-  element: Element,
-  slotName: string,
-  options?: GetSlottedOptions
-): (T | null) | T[] {
-  const slotSelector = `[slot="${slotName}"]`;
-
-  if (options?.all) {
-    return queryMultiple<T>(element, slotSelector, options);
-  }
-
-  return querySingle<T>(element, slotSelector, options);
-}
-
-function queryMultiple<T extends Element = Element>(
-  element: Element,
-  slotSelector: string,
-  options?: GetSlottedOptions
-): T[] {
-  let matches = Array.from(element.querySelectorAll<T>(slotSelector));
-  matches = options && options.direct === false ? matches : matches.filter((el) => el.parentElement === element);
-
-  const selector = options?.selector;
-  return selector ? matches.map((item) => item.querySelector<T>(selector)).filter((match) => !!match) : matches;
-}
-
-function querySingle<T extends Element = Element>(
-  element: Element,
-  slotSelector: string,
-  options?: GetSlottedOptions
-): T | null {
-  let match = element.querySelector<T>(slotSelector);
-  match = options && options.direct === false ? match : match?.parentElement === element ? match : null;
-
-  const selector = options?.selector;
-  return selector ? match.querySelector<T>(selector) : match;
-}
