--- src/components/utils/resources.ts
@@ -1,3 +1,0 @@
-export const CSS_UTILITY = {
-  rtl: "calcite--rtl"
-};
